# Conflicts:
#	.env.example
#	apps/products-service/tsconfig.app.json
#	apps/products-service/tsconfig.json

+++ apps/orders-service/prisma/migrations/20250924150755_update_orders_schema/migration.sql
@@ -2,7 +2,7 @@
 CREATE TABLE "Order" (
     "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
     "customerName" TEXT NOT NULL,
-    "totalAmount" INTEGER NOT NULL,
+    "totalAmount" DECIMAL NOT NULL,
     "status" TEXT NOT NULL DEFAULT 'pending',
     "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" DATETIME NOT NULL
@@ -15,7 +15,7 @@ CREATE TABLE "OrderItem" (
     "productId" INTEGER NOT NULL,
     "productName" TEXT NOT NULL,
     "quantity" INTEGER NOT NULL,
-    "price" INTEGER NOT NULL,
+    "price" DECIMAL NOT NULL,
     CONSTRAINT "OrderItem_orderId_fkey" FOREIGN KEY ("orderId") REFERENCES "Order" ("id") ON DELETE CASCADE ON UPDATE CASCADE
 );
